Move summaries into wallet package
